Working version of conda env, and upgraded stackgan code
note that the stackgan code is still listed as a git submodule just so it's clear
that we mostly did not write it; however, I've changed it a bunch from the original,
by upgrading the code to newer tensorflow version and changing some misc stuff.

I will make a copy of this in the next commit that we can use to make our modified
style-conditioned GAN, which will include these changes.